Moved binarySearch and getRlpHeaders from utils to common.helpers with tests and fixed wrapper functions to V2+ objects and functions
